# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Bootcamp Management System APIs",
    "version": "1.0",
    "description": "Bootcamp Management System APIs",
    "contact": {
      "name": "VMware "
    }
  },
  "servers": [
    {
      "url": "https://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/meetings": {
      "post": {
        "summary": "Post a new meeting",
        "tags": [
          "MeetingAPIs"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "meetObject": {
                      "type": "object",
                      "properties": {
                        "zoomlink": {
                          "type": "string"
                        },
                        "meetTitle": {
                          "type": "string"
                        },
                        "meetDescription": {
                          "type": "string"
                        },
                        "isRecorded": {
                          "type": "boolean"
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "message": "Posted a recorded meeting",
                      "meetObject": {
                        "zoomlink": "http://www.googe.com",
                        "meetTitle": "samplemeet2",
                        "meetDescription": "sample",
                        "isRecorded": true
                      }
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "message": "Posted a recorded meeting",
                      "meetObject": {
                        "zoomlink": "http://www.googe.com",
                        "meetTitle": "samplemeet2",
                        "meetDescription": "sample",
                        "isRecorded": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "operationId": "post-api-meetings",
        "description": "Post a new meeting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "zoomlink": {
                    "type": "string"
                  },
                  "meetTitle": {
                    "type": "string"
                  },
                  "meetDescription": {
                    "type": "string"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "endTime": {
                    "type": "string"
                  },
                  "isRecorded": {
                    "type": "boolean"
                  }
                },
                "x-examples": {
                  "example-1": {
                    "zoomlink": "http://www.google.com",
                    "meetTitle": "samplemeet",
                    "meetDescription": "sample",
                    "startTime": "2022/07/30/18:30",
                    "endTime": "2022/07/30/19:30",
                    "isRecorded": false
                  }
                }
              },
              "examples": {
                "Example": {
                  "value": {
                    "zoomlink": "http://www.google.com",
                    "meetTitle": "samplemeet",
                    "meetDescription": "sample",
                    "startTime": "2022/07/30/18:30",
                    "endTime": "2022/07/30/19:30",
                    "isRecorded": true
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all meetings",
        "operationId": "get-api-meetings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "meetingList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "zoomlink": {
                            "type": "string"
                          },
                          "meetTitle": {
                            "type": "string"
                          },
                          "meetDescription": {
                            "type": "string"
                          },
                          "startTime": {
                            "type": "string"
                          },
                          "endTime": {
                            "type": "string"
                          },
                          "isRecorded": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  },
                  "x-examples": {
                    "example-1": {
                      "message": "Get request to the only live meeting",
                      "meetingList": [
                        {
                          "_id": "62e18ac416994a532e55fd74",
                          "zoomlink": "http://www.google.com",
                          "meetTitle": "samplemeet",
                          "meetDescription": "sample",
                          "startTime": "2022-07-30T13:00:00.000Z",
                          "endTime": "2022-07-30T14:00:00.000Z",
                          "isRecorded": false
                        }
                      ]
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "message": "string",
                      "meetingList": [
                        {
                          "_id": "string",
                          "zoomlink": "string",
                          "meetTitle": "string",
                          "meetDescription": "string",
                          "startTime": "string",
                          "endTime": "string",
                          "isRecorded": true
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "live",
            "description": "Live meeting"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "recorded",
            "description": "Recorded meetings"
          }
        ],
        "tags": [
          "MeetingAPIs"
        ],
        "description": "Get all meetings"
      }
    },
...
  }
}